fix: update routing and text for modal
